requirements.txt
streamlit>=1.20
google-generativeai>=0.3.2
python-dotenv
